# pkg: support generic for btree