--- a/pkg/btree/.travis.yml
+++ b/pkg/btree/.travis.yml
@@ -1,1 +1,0 @@
-language: go