--- a/pkg/btree/btree_generic.go
+++ b/pkg/btree/btree_generic.go
@@ -73,7 +73,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// nolint
+//revive:disable
 package btree
 
 import (