--- a/server/statistics/buckets/hot_bucket_cache.go
+++ b/server/statistics/buckets/hot_bucket_cache.go
@@ -23,7 +23,7 @@ import (
 	"github.com/pingcap/log"
 	"github.com/tikv/pd/pkg/keyutil"
 	"github.com/tikv/pd/pkg/logutil"
-	"github.com/tikv/pd/pkg/rangetree"
+	"github.com/tikv/pd/server/core/rangetree"
 	"go.uber.org/zap"
 )